+++ tests/crud/create_order.rs
@@ -103,7 +103,7 @@ pub async fn test_create_order(db: &DbConn) {
     let order_model = order.unwrap();
     assert_eq!(order_model.total, dec!(15.10));
 
-    let customer: Option<customer::Model> = Customer::find_by_id(order_model.customer_id as u64)
+    let customer: Option<customer::Model> = Customer::find_by_id(order_model.customer_id)
         .one(db)
         .await
         .expect("could not find customer");
@@ -111,7 +111,7 @@ pub async fn test_create_order(db: &DbConn) {
     let customer_model = customer.unwrap();
     assert_eq!(customer_model.name, "Kate");
 
-    let bakery: Option<bakery::Model> = Bakery::find_by_id(order_model.bakery_id as i64)
+    let bakery: Option<bakery::Model> = Bakery::find_by_id(order_model.bakery_id)
         .one(db)
         .await
         .expect("could not find bakery");
